+++ test/hardhat/viem/pcsv3-automan-transaction.test.ts
@@ -4,7 +4,6 @@ import { AutomatedMarketMakerEnum } from 'aperture-lens/dist/src/viem';
 import { BigNumber } from 'ethers';
 import { defaultAbiCoder } from 'ethers/lib/utils';
 import hre, { ethers } from 'hardhat';
-import JSBI from 'jsbi';
 import {
   Address,
   PublicClient,
@@ -289,172 +288,6 @@ describe('Viem - PCSV3Automan transaction tests', function () {
     });
   });
 
-  it('Rebalance with 1inch', async function () {
-    const existingPosition = await PositionDetails.fromPositionId(
-      chainId,
-      amm,
-      positionId,
-      publicClient,
-    );
-    const { swapData, liquidity } = (
-      await getRebalanceSwapInfo(
-        chainId,
-        amm,
-        eoa,
-        positionId,
-        240000,
-        300000,
-        0.01 /*slippageTolerance*/,
-        ['60000', '3000'],
-        publicClient,
-        [E_Solver.OneInch],
-        existingPosition,
-        undefined /**blockNumber */,
-        false /** feesOn */,
-      )
-    )[0];
-    const { tx: txRequest } = await getRebalanceTx(
-      chainId,
-      amm,
-      eoa,
-      positionId,
-      240000,
-      300000,
-      /*slippageTolerance=*/ new Percent(1, 100),
-      /*deadlineEpochSeconds=*/ BigInt(Math.floor(Date.now() / 1000)),
-      publicClient,
-      swapData,
-      liquidity,
-      0n /** feeBips */,
-      existingPosition.position,
-    );
-    // Owner of position id 4 sets Automan as operator.
-    await testClient.impersonateAccount({ address: eoa });
-    const walletClient = testClient.extend(walletActions);
-    const txHash = await walletClient.sendTransaction({
-      to: txRequest.to,
-      data: txRequest.data,
-      account: txRequest.from,
-      // from: txRequest.from,
-      chain: walletClient.chain,
-    });
-    const txReceipt = await publicClient.getTransactionReceipt({
-      hash: txHash,
-    });
-    const newPositionId = getMintedPositionIdFromTxReceipt(
-      chainId,
-      amm,
-      txReceipt,
-      eoa,
-    )!;
-    expect(
-      await getBasicPositionInfo(chainId, amm, newPositionId, publicClient),
-    ).to.deep.contains({
-      token0: existingPosition.pool.token0,
-      token1: existingPosition.pool.token1,
-      fee: existingPosition.pool.fee,
-      tickLower: 240000,
-      tickUpper: 300000,
-    });
-  });
-
-  // Unit test with 1inch is known to be unstable, skip it for now.
-  it.skip('Optimal mint with 1inch', async function () {
-    const pool = await getPool(
-      WBTC_ADDRESS,
-      WETH_ADDRESS,
-      FeeAmount.PCS_V3_MEDIUM,
-      chainId,
-      amm,
-      publicClient,
-    );
-    const token0Amount = CurrencyAmount.fromRawAmount(
-      pool.token0,
-      '1000000000',
-    );
-    const token1Amount = CurrencyAmount.fromRawAmount(
-      pool.token1,
-      '1000000000000000000',
-    );
-    await dealERC20(
-      pool.token0.address as Address,
-      pool.token1.address as Address,
-      BigInt(token0Amount.quotient.toString()),
-      BigInt(token1Amount.quotient.toString()),
-      eoa,
-      getAMMInfo(chainId, amm)!.apertureAutoman,
-    );
-    const tickLower = nearestUsableTick(
-      pool.tickCurrent - 1000,
-      pool.tickSpacing,
-    );
-    const tickUpper = nearestUsableTick(
-      pool.tickCurrent + 1000,
-      pool.tickSpacing,
-    );
-    const { swapData, liquidity } = (
-      await getOptimalMintSwapInfo(
-        chainId,
-        amm,
-        token0Amount,
-        token1Amount,
-        pool.fee,
-        tickLower,
-        tickUpper,
-        eoa,
-        0.5,
-        publicClient,
-        [E_Solver.OneInch],
-      )
-    )[0];
-    const { tx: txRequest } = await getOptimalMintTx(
-      chainId,
-      amm,
-      token0Amount,
-      token1Amount,
-      pool.fee,
-      tickLower,
-      tickUpper,
-      eoa,
-      /*deadlineEpochSeconds=*/ BigInt(Math.floor(Date.now() / 1000)),
-      0.5,
-      publicClient,
-      swapData,
-      liquidity,
-    );
-    await testClient.impersonateAccount({ address: eoa });
-    const walletClient = testClient.extend(walletActions);
-    const txHash = await walletClient.sendTransaction({
-      to: txRequest.to,
-      data: txRequest.data,
-      account: txRequest.from,
-      // from: txRequest.from,
-      chain: walletClient.chain,
-    });
-    const txReceipt = await publicClient.getTransactionReceipt({
-      hash: txHash,
-    });
-    const newPositionId = getMintedPositionIdFromTxReceipt(
-      chainId,
-      amm,
-      txReceipt,
-      eoa,
-    )!;
-    const newPosition = await getBasicPositionInfo(
-      chainId,
-      amm,
-      newPositionId,
-      publicClient,
-    );
-    expect(newPosition).to.deep.contains({
-      token0: pool.token0,
-      token1: pool.token1,
-      fee: pool.fee,
-      tickLower,
-      tickUpper,
-    });
-  });
-
   it('Optimal mint without 1inch', async function () {
     const pool = await getPool(
       WBTC_ADDRESS,
@@ -552,96 +385,6 @@ describe('Viem - PCSV3Automan transaction tests', function () {
     });
   });
 
-  // Unit test with 1inch is known to be unstable, skip it for now.
-  it.skip('Increase liquidity optimal with 1inch', async function () {
-    const existingPosition = await PositionDetails.fromPositionId(
-      chainId,
-      amm,
-      positionId,
-      publicClient,
-    );
-    const pool = existingPosition.pool;
-    const token0Amount = CurrencyAmount.fromRawAmount(
-      pool.token0,
-      '1000000000',
-    );
-    const token1Amount = CurrencyAmount.fromRawAmount(
-      pool.token1,
-      '1000000000000000000',
-    );
-    await dealERC20(
-      pool.token0.address as Address,
-      pool.token1.address as Address,
-      BigInt(token0Amount.quotient.toString()),
-      BigInt(token1Amount.quotient.toString()),
-      eoa,
-      getAMMInfo(chainId, amm)!.apertureAutoman,
-    );
-    const { swapData, liquidity } = (
-      await getIncreaseLiquidityOptimalSwapInfo(
-        {
-          tokenId: Number(positionId),
-          slippageTolerance: new Percent(5, 1000),
-          deadline: Math.floor(Date.now() / 1000 + 60 * 30),
-        },
-        chainId,
-        amm,
-        token0Amount,
-        token1Amount,
-        eoa,
-        publicClient,
-        [E_Solver.OneInch],
-        existingPosition.position,
-      )
-    )[0];
-    const { tx: txRequest } = await getIncreaseLiquidityOptimalTx(
-      {
-        tokenId: Number(positionId),
-        slippageTolerance: new Percent(50, 100),
-        deadline: Math.floor(Date.now() / 1000 + 60 * 30),
-      },
-      chainId,
-      amm,
-      token0Amount,
-      token1Amount,
-      eoa,
-      publicClient,
-      swapData,
-      liquidity,
-      existingPosition.position,
-    );
-
-    // Owner of position id 4 sets Automan as operator.
-    await testClient.impersonateAccount({ address: eoa });
-    const walletClient = testClient.extend(walletActions);
-    const txHash = await walletClient.sendTransaction({
-      to: txRequest.to,
-      data: txRequest.data,
-      account: txRequest.from,
-      // from: txRequest.from,
-      chain: walletClient.chain,
-    });
-    await publicClient.getTransactionReceipt({
-      hash: txHash,
-    });
-    const newPosition = await getBasicPositionInfo(
-      chainId,
-      amm,
-      positionId,
-      publicClient,
-    );
-    expect(newPosition).to.deep.contains({
-      token0: pool.token0,
-      token1: pool.token1,
-      fee: pool.fee,
-      tickLower: existingPosition.tickLower,
-      tickUpper: existingPosition.tickUpper,
-    });
-    expect(
-      JSBI.GT(newPosition.liquidity!, existingPosition.liquidity!),
-    ).to.equal(true);
-  });
-
   it('Increase liquidity optimal without 1inch', async function () {
     const existingPosition = await PositionDetails.fromPositionId(
       chainId,
@@ -727,4 +470,6 @@ describe('Viem - PCSV3Automan transaction tests', function () {
       tickUpper: existingPosition.tickUpper,
     });
   });
+
+  // TODO: Add tests for rebalance, mintOptimal and increaseLiquidityOptimal with 1inch.
 });
